requirements.txt
django
djangorestframework
mysqlclient
python-dotenv
gunicorn
django-cors-headers  # 👈 Asegura que esta línea esté en el archivo
